Revert "bpf/Makefile: Forbid implicit int conversions"
This reverts commit ce6f2c7729df96f31a0637c6a24f24add3a65fa6.

+++ bpf/lib/conntrack.h
@@ -1024,7 +1024,7 @@ static __always_inline bool
 ct_has_nodeport_egress_entry4(const void *map,
 			      struct ipv4_ct_tuple *ingress_tuple)
 {
-	__u8 prev_flags = ingress_tuple->flags;
+	int prev_flags = ingress_tuple->flags;
 	struct ct_entry *entry;
 
 	ingress_tuple->flags = TUPLE_F_OUT;
@@ -1041,7 +1041,7 @@ static __always_inline bool
 ct_has_nodeport_egress_entry6(const void *map,
 			      struct ipv6_ct_tuple *ingress_tuple)
 {
-	__u8 prev_flags = ingress_tuple->flags;
+	int prev_flags = ingress_tuple->flags;
 	struct ct_entry *entry;
 
 	ingress_tuple->flags = TUPLE_F_OUT;
